Creating route handlers for task detail tabs
